copy: discount landing page
also deduplicated discount form field code
