+++ src/components/Form/CreateDiscountForm/CreateDiscountForm.tsx
@@ -8,24 +8,22 @@ import { CreateDiscount } from "../../../api/generated";
 import { Severity, useTooltip } from "../../../context/tooltip";
 import { DASHBOARD } from "../../../navigation/routes";
 import { RootState } from "../../../store/store";
-import Bucket from "../CreateProfileForm/DiscountData/Bucket";
-import DiscountConditions from "../CreateProfileForm/DiscountData/DiscountConditions";
-import DiscountInfo from "../CreateProfileForm/DiscountData/DiscountInfo";
-import DiscountUrl from "../CreateProfileForm/DiscountData/DiscountUrl";
-import EnrollToEyca from "../CreateProfileForm/DiscountData/EnrollToEyca";
-import LandingPage from "../CreateProfileForm/DiscountData/LandingPage";
-import ProductCategories from "../CreateProfileForm/DiscountData/ProductCategories";
-import StaticCode from "../CreateProfileForm/DiscountData/StaticCode";
-import FormField from "../FormField";
+import DiscountInfo, {
+  getDiscountTypeChecks
+} from "../CreateProfileForm/DiscountData/DiscountInfo";
 import FormSection from "../FormSection";
 import { discountDataValidationSchema } from "../ValidationSchemas";
-import { MAX_SELECTABLE_CATEGORIES } from "../../../utils/constants";
 import {
   discountEmptyInitialValues,
-  getDiscountTypeChecks,
   sanitizeDiscountFormValues
 } from "../EditDiscountForm/EditDiscountForm";
 
+/**
+ * These are the entry points for forms for discounts. This comment is repeated in every file.
+ * src/components/Form/CreateProfileForm/DiscountData/DiscountData.tsx Used in onboarding process
+ * src/components/Form/CreateDiscountForm/CreateDiscountForm.tsx Used to create new discount once onboarded
+ * src/components/Form/EditDiscountForm/EditDiscountForm.tsx  Used to edit new discount once onboarded
+ */
 const CreateDiscountForm = () => {
   const history = useHistory();
   const agreement = useSelector((state: RootState) => state.agreement.value);
@@ -36,24 +34,21 @@ const CreateDiscountForm = () => {
   });
   const profile = profileQuery.data;
 
-  const { checkStaticCode, checkLanding, checkBucket } = getDiscountTypeChecks(
-    profile
-  );
+  const { checkStaticCode, checkLanding, checkBucket } =
+    getDiscountTypeChecks(profile);
 
-  const createDiscountMutation = remoteData.Index.Discount.createDiscount.useMutation(
-    {
+  const createDiscountMutation =
+    remoteData.Index.Discount.createDiscount.useMutation({
       onSuccess() {
         history.push(DASHBOARD);
       },
       onError() {
         triggerTooltip({
           severity: Severity.DANGER,
-          text:
-            "Errore durante la creazione dell'opportunità, controllare i dati e riprovare"
+          text: "Errore durante la creazione dell'opportunità, controllare i dati e riprovare"
         });
       }
-    }
-  );
+    });
 
   const createDiscount = async (
     agreementId: string,
@@ -74,78 +69,11 @@ const CreateDiscountForm = () => {
       {({ values, setFieldValue, isSubmitting }) => (
         <Form autoComplete="off">
           <FormSection hasIntroduction>
-            <DiscountInfo formValues={values} setFieldValue={setFieldValue} />
-            <FormField
-              htmlFor="productCategories"
-              isTitleHeading
-              title="Categorie merceologiche"
-              description={`Seleziona al massimo ${MAX_SELECTABLE_CATEGORIES} categorie merceologiche a cui appatengono i beni/servizi oggetto dell'opportunità`}
-              isVisible
-              required
-            >
-              <ProductCategories
-                selectedCategories={values.productCategories}
-              />
-            </FormField>
-            <FormField
-              htmlFor="discountConditions"
-              isTitleHeading
-              title="Condizioni dell’opportunità"
-              description="Descrivere eventuali limitazioni relative all'opportunità (es. sconto valido per l’acquisto di un solo abbonamento alla stagione di prosa presso gli sportelli del teatro) - Max 200 caratteri"
-              isVisible
-            >
-              <DiscountConditions />
-            </FormField>
-            {!checkLanding && (
-              <FormField
-                htmlFor="discountUrl"
-                title="Link all'opportunità"
-                description="Inserire l’URL di destinazione del sito o dell’app da cui i titolari di CGN potranno accedere all'opportunità"
-                isTitleHeading
-                isVisible
-              >
-                <DiscountUrl />
-              </FormField>
-            )}
-            {checkStaticCode && (
-              <FormField
-                htmlFor="staticCode"
-                isTitleHeading
-                title="Codice statico"
-                description="Inserire il codice relativo all'opportunità che l’utente dovrà inserire sul vostro portale online"
-                isVisible
-                required
-              >
-                <StaticCode />
-              </FormField>
-            )}
-            {checkLanding && (
-              <FormField
-                htmlFor="landingPage"
-                isTitleHeading
-                title="Indirizzo della landing page*"
-                description="Inserire l’URL della landing page da cui i titolari di CGN potranno accedere all’opportunità"
-                isVisible
-                required
-              >
-                <LandingPage />
-              </FormField>
-            )}
-            {checkBucket && (
-              <Bucket
-                agreementId={agreement.id}
-                label={"Seleziona un file dal computer"}
-                formValues={values}
-                setFieldValue={setFieldValue}
-              />
-            )}
-            {profile && (
-              <EnrollToEyca
-                profile={profile}
-                formValues={values}
-                setFieldValue={setFieldValue}
-              />
-            )}
+            <DiscountInfo
+              formValues={values}
+              setFieldValue={setFieldValue}
+              profile={profile}
+            />
             <div className="mt-10">
               <Button
                 className="px-14 mr-4"
