+++ src/components/Form/CreateProfileForm/DiscountData/DiscountInfo.tsx
@@ -3,17 +3,31 @@
 import { Field, FieldProps } from "formik";
 import React from "react";
 import DatePicker from "react-datepicker";
+import { useSelector } from "react-redux";
 import CustomErrorMessage from "../../CustomErrorMessage";
 import DateInputComponent from "../../DateInputComponent";
 import InputField from "../../FormField";
+import FormField from "../../FormField";
+import { MAX_SELECTABLE_CATEGORIES } from "../../../../utils/constants";
+import { Profile } from "../../../../api/generated";
+import { RootState } from "../../../../store/store";
+import ProductCategories from "./ProductCategories";
+import DiscountConditions from "./DiscountConditions";
+import EnrollToEyca from "./EnrollToEyca";
+import Bucket from "./Bucket";
+import LandingPage from "./LandingPage";
+import StaticCode from "./StaticCode";
+import DiscountUrl from "./DiscountUrl";
 
 type Props = {
   formValues?: any;
   setFieldValue?: any;
   index?: number;
+  profile: Profile | undefined;
 };
 
-const DiscountInfo = ({ formValues, setFieldValue, index }: Props) => {
+// eslint-disable-next-line complexity
+const DiscountInfo = ({ formValues, setFieldValue, index, profile }: Props) => {
   const hasIndex = index !== undefined;
 
   const dateFrom = hasIndex
@@ -23,6 +37,11 @@ const DiscountInfo = ({ formValues, setFieldValue, index }: Props) => {
   const dateTo = hasIndex
     ? formValues.discounts[index as number].endDate
     : formValues.endDate;
+
+  const { checkBucket, checkLanding, checkStaticCode } =
+    getDiscountTypeChecks(profile);
+
+  const agreement = useSelector((state: RootState) => state.agreement.value);
 
   return (
     <>
@@ -67,7 +86,7 @@ const DiscountInfo = ({ formValues, setFieldValue, index }: Props) => {
       <InputField
         htmlFor="description"
         title="Descrizione opportunità"
-        description="Se necessario, inserire una descrizione più approfondita dell'opportunità (es. Sconto valido per l'acquisto di due ingressi alla stagione di prosa 2021/22 presso il Teatro Comunale) - Max 250 caratteri"
+        description="Se necessario, inserire una descrizione più approfondita dell’opportunità (es. Sconto valido per l’acquisto di due ingressi alla stagione di prosa 2021/22 presso il Teatro Comunale) - Max 250 caratteri"
         isVisible
       >
         <div className="row">
@@ -121,8 +140,8 @@ const DiscountInfo = ({ formValues, setFieldValue, index }: Props) => {
         <div className="col-5">
           <InputField
             htmlFor="startDate"
-            title="Data di inizio dell'opportunità"
-            description="Indicare il giorno e l’ora da cui l'opportunità diventa valida"
+            title="Data di inizio dell’opportunità"
+            description="Indicare il giorno e l’ora da cui l’opportunità diventa valida"
             isVisible
             required
           >
@@ -186,7 +205,7 @@ const DiscountInfo = ({ formValues, setFieldValue, index }: Props) => {
       <InputField
         htmlFor="discount"
         title="Entità dello sconto"
-        description="Se l'opportunità lo prevede, inserire la percentuale (%) di sconto erogata"
+        description="Se l’opportunità lo prevede, inserire la percentuale (%) di sconto erogata"
         isVisible
       >
         <Field
@@ -219,8 +238,108 @@ const DiscountInfo = ({ formValues, setFieldValue, index }: Props) => {
           name={hasIndex ? `discounts[${index}].discount` : "discount"}
         />
       </InputField>
+      <FormField
+        htmlFor="productCategories"
+        isTitleHeading
+        title="Categorie merceologiche"
+        description={`Seleziona al massimo ${MAX_SELECTABLE_CATEGORIES} categorie merceologiche a cui appatengono i beni/servizi oggetto dell’opportunità`}
+        isVisible
+        required
+      >
+        <ProductCategories
+          selectedCategories={formValues.productCategories}
+          index={index}
+        />
+      </FormField>
+      <FormField
+        htmlFor="discountConditions"
+        isTitleHeading
+        title="Condizioni dell’opportunità"
+        description="Descrivere eventuali limitazioni relative all’opportunità (es. sconto valido per l’acquisto di un solo abbonamento alla stagione di prosa presso gli sportelli del teatro) - Max 200 caratteri"
+        isVisible
+      >
+        <DiscountConditions index={index} />
+      </FormField>
+      {!checkLanding && (
+        <FormField
+          htmlFor="discountUrl"
+          title="Link all’opportunità"
+          description="Inserire l’URL di destinazione del sito o dell’app da cui i titolari di CGN potranno accedere all’opportunità"
+          isTitleHeading
+          isVisible
+        >
+          <DiscountUrl index={index} />
+        </FormField>
+      )}
+      {checkStaticCode && (
+        <FormField
+          htmlFor="staticCode"
+          isTitleHeading
+          title="Codice statico"
+          description="Inserire il codice relativo all’opportunità che l’utente dovrà inserire sul vostro portale online"
+          isVisible
+          required
+        >
+          <StaticCode index={index} />
+        </FormField>
+      )}
+      {checkLanding && (
+        <FormField
+          htmlFor="landingPage"
+          isTitleHeading
+          title="Indirizzo della landing page*"
+          description="Inserisci l’URL della pagina web da cui sarà possibile accedere all’opportunità"
+          isVisible
+          required
+        >
+          <LandingPage index={index} />
+        </FormField>
+      )}
+      {checkBucket && (
+        <Bucket
+          agreementId={agreement.id}
+          label={"Seleziona un file dal computer"}
+          index={index}
+          formValues={formValues}
+          setFieldValue={setFieldValue}
+        />
+      )}
+      {profile && (
+        <EnrollToEyca
+          profile={profile}
+          index={index}
+          formValues={formValues}
+          setFieldValue={setFieldValue}
+        />
+      )}
     </>
   );
 };
 
 export default DiscountInfo;
+
+export function getDiscountTypeChecks(profile: Profile | undefined) {
+  const onlineOrBoth =
+    profile?.salesChannel?.channelType === "OnlineChannel" ||
+    profile?.salesChannel?.channelType === "BothChannels";
+
+  const checkStaticCode =
+    onlineOrBoth &&
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore
+    profile?.salesChannel?.discountCodeType === "Static";
+
+  const checkLanding =
+    onlineOrBoth &&
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore
+    profile?.salesChannel?.discountCodeType === "LandingPage";
+
+  const checkBucket =
+    onlineOrBoth &&
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore
+    profile?.salesChannel?.discountCodeType === "Bucket";
+
+  return { checkStaticCode, checkLanding, checkBucket };
+}
